move exportOptionsSchema to /lib/network-exporters/utils
