+++ app/(dashboard)/dashboard/interviews/_components/ExportOptionsView.tsx
@@ -1,6 +1,6 @@
 import { type Dispatch, type SetStateAction } from 'react';
 import { Switch } from '~/components/ui/switch';
-import { type ExportOptions } from '~/utils/validateExportOptions';
+import { type ExportOptions } from '~/lib/network-exporters/utils/exportOptionsSchema';
 
 type ExportOptionsViewProps = {
   exportOptions: ExportOptions;
